// src/shared/data/regex/vendor/waystones/Waystone.Gen.ts
export interface WaystoneRegex {
  id: number
  name: string
  regex: string
  values: number[]
  ranges: number[][]
  prefix: boolean
  tags: string[]
}